requirements.txt
Django==3.1.7
six==1.15.0
django-taggit==1.3.0
django-reset-migrations==0.4.0